refactor: streamline instance retrieval and improve code consistency

+++ src/Utils/Utils.java
@@ -33,7 +33,7 @@ public class Utils {
     }
 
     public static double easeInOutCubic(double t) {
-        return t < 0.5 ? 4 * t * t * t : 1 - (double) Math.pow(-2 * t + 2, 3) / 2;
+        return t < 0.5 ? 4 * t * t * t : 1 - Math.pow(-2 * t + 2, 3) / 2;
     }
 
 }
